Move comparator tests into separate suite
